Merge bot and dashboard to a monorepo using Turborepo
This commit does a lot of things, I know I should pick and commit
things seperately but you know how it is.

Created this repo using hraban/tomono, merging aurene-bot and
aurene-dashboard repositories to a single repository while maintaining
the commit history. Then added Turborepo stuff on top of it.

The main goal here was to share the database between bot and dashboard,
so with this commit, database is seperated to it's own package under
/packages and then accessed by the bot and dashboard. Pretty cool!